+++ features/Context/NavigationContext.php
@@ -633,7 +633,9 @@ class NavigationContext extends RawMinkContext implements PageObjectAwareInterfa
     {
         $expectedAddress = $this->getPage('Product edit')->getUrl(['id' => $product->getId()]);
         $this->assertAddress($expectedAddress);
-        $this->wait();
+        $this->getMainContext()->spin(function () {
+            return $this->getCurrentPage()->find('css', '.product-label');
+        });
     }
 
     /**