+++ src/Pim/Bundle/EnrichBundle/Resources/public/js/product/form/panel/completeness.js
@@ -21,7 +21,6 @@ define(
         return BaseForm.extend({
             template: _.template(template),
             className: 'panel-pane',
-            code: 'completeness',
             events: {
                 'click header': 'switchLocale',
                 'click .missing-attributes span': 'showAttribute'
@@ -36,10 +35,8 @@ define(
                     label: _.__('pim_enrich.form.product.panel.completeness.title')
                 });
 
-                mediator.on('product:action:post_update', _.bind(this.update, this));
-
-                this.stopListening(mediator, 'change-family:change:after');
-                this.listenTo(mediator, 'change-family:change:after', _.bind(this.onChangeFamily, this));
+                this.listenTo(mediator, 'pim_enrich:form:entity:post_update', this.update);
+                this.listenTo(mediator, 'pim_enrich:form:change-family:after', this.onChangeFamily);
 
                 return BaseForm.prototype.configure.apply(this, arguments);
             },
@@ -48,6 +45,10 @@ define(
              * {@inheritdoc}
              */
             render: function () {
+                if (this.code !== this.getParent().state.get('currentPanel')) {
+                    return this;
+                }
+
                 if (this.getFormData().meta) {
                     $.when(
                         FetcherRegistry.getFetcher('completeness').fetchForProduct(
@@ -92,7 +93,7 @@ define(
              */
             showAttribute: function (event) {
                 mediator.trigger(
-                    'show_attribute',
+                    'pim_enrich:form:show_attribute',
                     {
                         attribute: event.currentTarget.dataset.attribute,
                         locale: event.currentTarget.dataset.locale,